requirements.txt
arrow==1.3.0
asgiref==3.10.0
binaryornot==0.4.4
certifi==2025.10.5
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
cloudinary==1.44.1
colorama==0.4.6
cookiecutter==2.6.0
dj-database-url==3.0.1
Django==5.2
django-allauth==65.12.0
django-browser-reload==1.21.0
django-cloudinary-storage==0.3.0
django-tailwind==4.2.0
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.11
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
types-python-dateutil==2.9.0.20251008
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
